LayoutTests/imported/w3c/web-platform-tests/webrtc-extensions/RTCRtpSynchronizationSource-senderCaptureTimeOffset-expected.txt



PASS [audio] getSynchronizationSources() should not contain senderCaptureTimeOffset if absolute capture time RTP header extension is not offered
PASS [audio] getSynchronizationSources() should not contain senderCaptureTimeOffset if absolute capture time RTP header extension is offered, but not answered
FAIL [audio] getSynchronizationSources() should contain senderCaptureTimeOffset if absolute capture time RTP header extension is negotiated step_wait_func: No abs-capture-time capture time header extension. Timed out waiting on condition
PASS [video] getSynchronizationSources() should not contain senderCaptureTimeOffset if absolute capture time RTP header extension is not offered
PASS [video] getSynchronizationSources() should not contain senderCaptureTimeOffset if absolute capture time RTP header extension is offered, but not answered
FAIL [video] getSynchronizationSources() should contain senderCaptureTimeOffset if absolute capture time RTP header extension is negotiated step_wait_func: No abs-capture-time capture time header extension. Timed out waiting on condition
FAIL Audio and video RTCRtpSynchronizationSource.senderCaptureTimeOffset must be zero step_wait_func: No abs-capture-time capture time header extension. Timed out waiting on condition
